+++ slipp/src/main/java/slipp/SlippWebApplicationInitializer.java
@@ -5,6 +5,7 @@ import nextstep.mvc.handleradapter.HandlerAdapter;
 import nextstep.mvc.handlermapping.HandlerMapping;
 import nextstep.mvc.handlermapping.InOrderHandlerMapping;
 import nextstep.mvc.handlermapping.AnnotationHandlerMapping;
+import nextstep.mvc.handlermapping.ManualHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
